test: fix tests

+++ test2code/src/test/kotlin/com/epam/drill/test2code/coverage/CoverageRetentionTest.kt
@@ -37,7 +37,7 @@ class CoverageRetentionTest {
     @BeforeEach
     fun setUp() {
         sender = mock()
-        inMemoryRetentionQueue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(375))
+        inMemoryRetentionQueue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(375))
         coverageTransport = spy(WebsocketCoverageTransport(UUID.randomUUID().toString(), sender))
         coverageSender = IntervalCoverageSender(2500, coverageTransport, inMemoryRetentionQueue) {
             sequenceOf(
@@ -99,7 +99,7 @@ class CoverageRetentionTest {
 
     @Test
     fun `queue limit exceeded`() = runBlocking {
-        inMemoryRetentionQueue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
+        inMemoryRetentionQueue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
         coverageSender = IntervalCoverageSender(2500, coverageTransport, inMemoryRetentionQueue) {
             sequenceOf(
                 ExecDatum(
