+++ test2code/src/test/kotlin/com/epam/drill/test2code/coverage/InMemoryRetentionQueueTest.kt
@@ -25,7 +25,7 @@ class InMemoryRetentionQueueTest {
         val byteCount = 6
         val innerArraySize = 2
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
-        val sizeLimit = BigDecimal.valueOf(10)
+        val sizeLimit = BigInteger.valueOf(10)
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
 
         queue.addAll(bytes)
@@ -39,7 +39,7 @@ class InMemoryRetentionQueueTest {
         val byteCount = 1
         val innerArraySize = 10
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
-        val sizeLimit = BigDecimal.valueOf(10)
+        val sizeLimit = BigInteger.valueOf(10)
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
 
         queue.addAll(bytes)
@@ -50,7 +50,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `adding single element larger than limit should result in an empty queue`() {
-        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
+        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
         val byteCount = 1
         val innerArraySize = 12
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
@@ -62,7 +62,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `flush should return the same elements as added to the queue`() {
-        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
+        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
         val byteCount = 1
         val innerArraySize = 5
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
@@ -75,7 +75,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `queue after flush must be empty`() {
-        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
+        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
         val byteCount = 1
         val innerArraySize = 5
         val bytes = Array(byteCount) { ByteArray(innerArraySize) { 1 } }.asSequence()
@@ -90,7 +90,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `queue should only fit elements with size less than or equal to total limit`() {
-        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigDecimal.valueOf(10))
+        val queue = InMemoryRetentionQueue(totalSizeByteLimit = BigInteger.valueOf(10))
         val expectedSizeOfQueue = 4
         val bytes = sequenceOf(
             ByteArray(12) { 1 },
@@ -106,7 +106,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `queue should contain elements within limit and exclude element exceeding limit `() {
-        val sizeLimit = BigDecimal.valueOf(10)
+        val sizeLimit = BigInteger.valueOf(10)
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
         val expectedBytes1 = ByteArray(2) { 1 }
         val expectedBytes2 = ByteArray(2) { 2 }
@@ -131,7 +131,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `adding new elements after flush should be reflected in queue content`() {
-        val sizeLimit = BigDecimal.valueOf(10)
+        val sizeLimit = BigInteger.valueOf(10)
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
         val expectedBytes1 = ByteArray(2) { 1 }
         val expectedBytes2 = ByteArray(2) { 2 }
@@ -154,7 +154,7 @@ class InMemoryRetentionQueueTest {
 
     @Test
     fun `initializing queue with negative totalSizeByteLimit should throw exception`() {
-        val sizeLimit = BigDecimal.valueOf(-10)
+        val sizeLimit = BigInteger.valueOf(-10)
 
         assertThrows<IllegalArgumentException> {
             val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
@@ -164,7 +164,7 @@ class InMemoryRetentionQueueTest {
     @Test
     fun `adding elements to a queue with zero totalSizeByteLimit should return an empty queue`() {
         val expectedSizeOfQueue = 0
-        val sizeLimit = BigDecimal.ZERO
+        val sizeLimit = BigInteger.ZERO
         val queue = InMemoryRetentionQueue(totalSizeByteLimit = sizeLimit)
         val bytes = sequenceOf(ByteArray(1) { 1 })
